Move SnapshotExtensions and FluentAssertionsExtensions to correct place
