+++ src/Snapshooter/Extensions/FluentAssertionsExtension.cs
@@ -1,0 +1,17 @@
+// Extension of Fluent assertion to not loose type information of asserted object
+namespace FluentAssertions.Primitives;
+
+public class TypedAssertions<T> : ObjectAssertions<T, TypedAssertions<T>>
+{
+    public TypedAssertions(T value) : base(value)
+    {
+    }
+}
+
+public static class Extensions
+{
+    public static TypedAssertions<TSubject> Should<TSubject>(this TSubject actualValue)
+    {
+        return new TypedAssertions<TSubject>(actualValue);
+    }
+}
